UI: use 'preferences-system' icon and make it bigger
The 'configure' icon is not a standard icon theme icon and probably only
available in the Breeze theme, so the preferences button was basically
invisible for many users.

Fix this by using the 'preferences-system' icon; and also increase the
size of the icon from 16x16 to 20x20 to make it more visible.

Fixes #50