Don't call applyEffect with bogus parameter for Custom Effect
Previously when the "Custom Effect" combobox item was added and
selected, this callback was called but because there's no item data we
kind of just by conincidence did the correct things but also called the
callback with RazerEffect::Off which again, by conincidence didn't have
any bad side effects.

Still for predictable behavior, handle this case better and more
explicitly.

And for even more coincidence, this was avoiding a bug in OpenRazer that
didn't stop the ripple effect properly, this is fixed with
https://github.com/openrazer/openrazer/pull/2117 so we can also safely
directly call custom effect methods after having ripple enabled.