DpiStageWidget: Enable setting 1-step DPI via spinbox
Some users want to set the DPI to e.g. 645 DPI which is not possible
with the 50-step DPI slider (and we want to keep the slider simple and
not have even more steps), so let's use the spinbox to enable this.

Issue: https://github.com/z3ntu/RazerGenie/issues/168